requirements.txt
# Playwright Stealth Email Scraper - Dependencies

# Web Automation
playwright>=1.40.0
playwright-stealth>=1.0.6

# Async support
aiohttp>=3.9.0
aiofiles>=23.0.0

# HTML Parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data Processing
pandas>=2.1.0

# Configuration
python-dotenv>=1.0.0

# User Agent Generation (optional, we have built-in list)